Allow running AWS 2 tests both grouped and isolated
