+++ integration-tests-aws2/aws2-s3/src/main/resources/application.properties
