+++ integration-tests-aws2/aws2-sqs-sns/README.adoc
