+++ integration-tests-aws2/aws2-sqs-sns/src/main/resources/application.properties
