+++ integration-tests/aws2-grouped/README.adoc
@@ -1,0 +1,19 @@
+== AWS 2 grouped integration tests
+
+This module allows executing all tests from `integration-tests-aws2/*` grouped together.
+
+While grouping brings the benefit of faster native compilation
+(compared to the sum of compilation times of all modules under `integration-tests-aws2/*`)
+it is methodologically not as clean as running those tests in isolation.
+Hence, this module is intended to be run instead of the tests under `integration-tests-aws2/*`
+if, in the given situation, the speed matters more than isolation.
+
+=== How the grouping works
+
+The grouped module dynamically pulls all sources from the individual AWS test modules to its
+`target/[test-]classes` directories.
+`application.properties` files and service descriptors are concatenated using a Groovy script.
+
+=== Mock backend vs. real service
+
+Please refer to the README files in `integration-tests-aws2` for more details.
