--- integration-tests/aws2-sqs-sns/src/test/resources/META-INF/services/org.apache.camel.quarkus.test.support.aws2.Aws2TestEnvCustomizer
@@ -1,1 +1,0 @@
-org.apache.camel.quarkus.component.aws2.sqs.it.Aws2SqsTestEnvCustomizer
