+++ integration-tests/pom.xml
@@ -68,8 +68,7 @@
         <module>avro-rpc</module>
         <module>aws</module>
         <module>aws2</module>
-        <module>aws2-s3</module>
-        <module>aws2-sqs-sns</module>
+        <module>aws2-grouped</module>
         <module>azure</module>
         <module>azure-eventhubs</module>
         <module>azure-storage-blob</module>
