+++ pom.xml
@@ -201,6 +201,7 @@
         <module>catalog</module>
         <module>integration-tests-support</module>
         <module>integration-tests</module>
+        <module>integration-tests-aws2</module>
         <module>docs</module>
     </modules>
 
@@ -783,6 +784,7 @@
                                         <sortModulesPath>extensions/pom.xml</sortModulesPath>
                                         <sortModulesPath>extensions-jvm/pom.xml</sortModulesPath>
                                         <sortModulesPath>integration-tests/pom.xml</sortModulesPath>
+                                        <sortModulesPath>integration-tests-aws2/pom.xml</sortModulesPath>
                                     </sortModulesPaths>
                                     <sortDependencyManagementPaths>
                                         <sortDependencyManagementPath>poms/bom/pom.xml</sortDependencyManagementPath>
@@ -793,6 +795,7 @@
                                             <includes>
                                                 <include>extensions-jvm/*/integration-test/pom.xml</include>
                                                 <include>integration-tests/*/pom.xml</include>
+                                                <include>integration-tests-aws2/*/pom.xml</include>
                                             </includes>
                                         </updateVirtualDependency>
                                     </updateVirtualDependencies>
@@ -827,9 +830,10 @@
                                         </extensionDir>
                                     </extensionDirs>
                                     <removeEmptyApplicationProperties>
-                                        <directory>${basedir}/integration-tests</directory>
+                                        <directory>${basedir}</directory>
                                         <includes>
-                                            <include>*/src/main/resources/application.properties</include>
+                                            <include>integration-tests/*/src/main/resources/application.properties</include>
+                                            <include>integration-tests-aws2/*/src/main/resources/application.properties</include>
                                         </includes>
                                     </removeEmptyApplicationProperties>
                                 </configuration>
