+++ tooling/scripts/test-categories.yaml
@@ -166,8 +166,7 @@ platform:
   - rest
   - rest-openapi
   - csimple
-  - aws2-s3
-  - aws2-sqs-sns
+  - aws2-grouped
 saas:
   - box
   - github
